refactor: unify HiSparse pool mapping internals
`HiSparseC4DevicePool` and `HiSparseDSATokenToKVPool` each carried their own
copy of the full -> HiSparse device index mapping helpers (`register_mapping`,
`translate_loc_to_hisparse_device`, `_translate_loc_to_hisparse_device`,
`translate_loc_from_full_to_hisparse_device`). The bodies were identical apart
from the C4 pool's `.to(torch.int32)` cast.

Extract them into `HiSparseDevicePoolMappingMixin`
(`mem_cache/base_hisparse_memory_pool.py`) and mix it into both pools. The C4
pool keeps its int32 cast by overriding `translate_loc_to_hisparse_device` and
delegating to `super()`. `translate_loc_from_full_to_hisparse_device` now
always routes through `translate_loc_from_full_to_compressed`, which is the
identity for the DSA pool, so behaviour is unchanged for both.

No public API change: every call site keeps the same method names.
